CogVM source as per VMMaker.oscog-eem.1594
Eliminate warnings for %.*s string length parameters in printf.

Cogit:
Don't inline AndCq:R:R: because its translation contains a jump and hence
won't work in a statement list expression.

Nuke genJump[Not]SmallInteger:scratch: in favour of genJump[Not]SmallInteger:scratchReg:

Slang:
Eliminate leaves in statement lists generated as expressions (to reduce
warnings).


git-svn-id: http://squeakvm.org/svn/squeak/branches/Cog@3525 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: 4cd0732d5258b26216bede0e8a8f582079e1d036

+++ platforms/Cross/vm/dispdbg.h
@@ -38,14 +38,14 @@
 #elif SENDTRACE /* send tracing.  */
 # define sendBreakpointreceiver(sel, len, rcvr) do { \
 	if (sendTrace) \
-		printf("%.*s\n", len, (char *)(sel)); \
+		printf("%.*s\n", (int)(len), (char *)(sel)); \
 } while (0)
 # define mnuBreakpointreceiver(sel, len, rcvr) 0
 
 #elif 0 /* send trace/byte count.  */
 # define sendBreakpointreceiver(sel, len, rcvr) do { \
 	if (sendTrace) \
-		printf("%u %.*s\n", GIV(byteCount), len, (char *)(sel)); \
+		printf("%u %.*s\n", GIV(byteCount), (int)(len), (char *)(sel)); \
 } while (0)
 # define mnuBreakpointreceiver(sel, len, rcvr) 0
 
@@ -57,7 +57,7 @@
 		if (0) sendTrace = 1; \
 	} \
 	if (sendTrace) \
-		printf("%.*s\n", len, (char *)(sel)); \
+		printf("%.*s\n", (int)(len), (char *)(sel)); \
 } while (0)
 # define mnuBreakpointreceiver(sel, len, rcvr) do { \
 	if ((len) == -breakSelectorLength \
